No more custom symgroups
Only built-in "default" and "whitespace" symgroups will now exist. Declaring symgroups is no longer possible.

The reason for this is that custom symgroups seem to be rather useless, and replaceable by symindies in almost all cases. While this makes a function like '+=' consist of 2 tokens rather than 1, many functions consist of multiple tokens anyway (i.e. '$a could be $b') so it shouldn't matter.